requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
kombu==4.6.7
Pillow==7.0.0
python-dotenv==0.10.5
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
vine==1.3.0
vk-api==11.7.0
